Create FileConfigTrait to avoid duplicate code in file based config engines.
